requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
flask-sqlalchemy==3.0.3
flask-caching==2.0.2
dnspython==2.3.0
pyOpenSSL==23.1.1
requests==2.28.2
cryptography==39.0.2
python-dotenv==1.0.0
gunicorn==20.1.0
